ai/requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0

# AI & ML Dependencies
groq==0.3.0
requests==2.31.0
sentence-transformers==2.2.2
numpy==1.24.3
chromadb==0.4.15

# Data Generation
sdv==1.2.0
pandas==2.0.3
scikit-learn==1.3.2

# Utilities
python-dateutil==2.8.2
uuid==1.30
readline==6.2.4.1  # For Unix-based systems only

# Logging
structlog==23.2.0

# For development only
black==23.11.0
flake8==6.1.0
pytest==7.4.3


# Voice processing
SpeechRecognition==3.10.0
PyAudio==0.2.13
gTTS==2.3.2
pygame==2.5.2
pydub==0.25.1

# Audio processing
wave
pyaudio


# Voice processing
elevenlabs==0.2.27
SpeechRecognition==3.10.0
PyAudio==0.2.13
gTTS==2.3.2
pygame==2.5.2
redis>=4.5.0
hiredis>=2.0.0
python-dotenv>=1.0.0
aiofiles==23.2.1